DEV: Reorganize user and anonymous sidebar files into respective folders
This commit does not change any functionality
