+++ app/assets/javascripts/discourse/app/components/sidebar.hbs
@@ -1,9 +1,9 @@
 <DSection @pageClass="has-sidebar" @class="sidebar-container" @scrollTop={{false}}>
   <div class="sidebar-scroll-wrap">
     {{#if this.currentUser}}
-      <Sidebar::Sections @collapsableSections={{true}}/>
+      <Sidebar::User::Sections @collapsableSections={{true}}/>
     {{else}}
-      <Sidebar::AnonymousSections />
+      <Sidebar::Anonymous::Sections />
     {{/if}}
   </div>
