--- app/assets/javascripts/discourse/app/components/sidebar/anonymous-sections.js
@@ -1,3 +1,0 @@
-import Component from "@glimmer/component";
-
-export default class SidebarAnonymousSectuons extends Component {}
