+++ app/assets/javascripts/discourse/app/components/sidebar/anonymous/sections.hbs
