+++ app/assets/javascripts/discourse/app/components/sidebar/anonymous/sections.js
@@ -1,0 +1,3 @@
+import Component from "@glimmer/component";
+
+export default class SidebarAnonymousSections extends Component {}
