+++ app/assets/javascripts/discourse/app/components/sidebar/hamburger-dropdown.hbs
@@ -3,7 +3,7 @@
     <div class="panel-body">
       <div class="panel-body-contents">
         <div class="sidebar-hamburger-dropdown">
-          <Sidebar::Sections @collapsableSections={{false}}/>
+          <Sidebar::User::Sections @collapsableSections={{false}}/>
           <Sidebar::Footer @tagName="" />
         </div>
       </div>
