+++ app/assets/javascripts/discourse/app/components/sidebar/user/categories-section.hbs
