+++ app/assets/javascripts/discourse/app/components/sidebar/user/community-section.hbs
