+++ app/assets/javascripts/discourse/app/components/sidebar/user/messages-section.js
@@ -3,8 +3,8 @@ import { cached } from "@glimmer/tracking";
 import { getOwner } from "discourse-common/lib/get-owner";
 import Component from "@glimmer/component";
 import { bind } from "discourse-common/utils/decorators";
-import GroupMessageSectionLink from "discourse/lib/sidebar/messages-section/group-message-section-link";
-import PersonalMessageSectionLink from "discourse/lib/sidebar/messages-section/personal-message-section-link";
+import GroupMessageSectionLink from "discourse/lib/sidebar/user/messages-section/group-message-section-link";
+import PersonalMessageSectionLink from "discourse/lib/sidebar/user/messages-section/personal-message-section-link";
 import { inject as service } from "@ember/service";
 
 export const INBOX = "inbox";
@@ -23,7 +23,7 @@ export const PERSONAL_MESSAGES_INBOX_FILTERS = [
 
 export const GROUP_MESSAGES_INBOX_FILTERS = [INBOX, NEW, UNREAD, ARCHIVE];
 
-export default class SidebarMessagesSection extends Component {
+export default class SidebarUserMessagesSection extends Component {
   @service appEvents;
   @service pmTopicTrackingState;
   @service currentUser;
