+++ app/assets/javascripts/discourse/app/components/sidebar/user/tags-section.hbs
