+++ app/assets/javascripts/discourse/app/lib/sidebar/user/base-custom-sidebar-section-link.js
