+++ app/assets/javascripts/discourse/app/lib/sidebar/user/categories-section/category-section-link.js
