+++ app/assets/javascripts/discourse/app/lib/sidebar/user/community-section/base-section-link.js
