+++ app/assets/javascripts/discourse/app/lib/sidebar/user/community-section/users-section-link.js
@@ -1,6 +1,6 @@
 import I18n from "I18n";
 
-import BaseSectionLink from "discourse/lib/sidebar/community-section/base-section-link";
+import BaseSectionLink from "discourse/lib/sidebar/user/community-section/base-section-link";
 
 export default class UsersSectionLink extends BaseSectionLink {
   get name() {
