+++ app/assets/javascripts/discourse/app/lib/sidebar/user/tags-section/pm-tag-section-link.js
