+++ app/assets/javascripts/discourse/app/lib/sidebar/user/tags-section/tag-section-link.js
